Modernised the `Tags` component
- Moved `Tags` and corresponding tests out of `shared` directory
- Replaced `shared.stories.tsx` with `tags.stories.tsx` and updated it
- Refactored `tags.tsx` to modularise styles and rely less on structure
- Tidied up the tests file
